move preparation files
