+++ scripts/data/transformers/magnitude-transformer.js
