+++ scripts/data/validators/schema-validator.js
